[cli] split terraform generate functions into module files
